requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.37.0
h5py==3.2.1
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.8
pydub==0.25.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
